refactor(lessons): LessonEditor sous 300 lignes — éclatement complet (H4)
989→199 lignes. Lève la déviation laissée en PR #86 : les sections du formulaire
sont désormais éclatées en sous-composants présentationnels (sans partial SCSS,
choix utilisateur → chrome de formulaire dupliqué VERBATIM dans chaque section) :
- LessonBasicInfoFields (208), LessonContentTypePicker (105),
  LessonContentFields (256) + LessonRichTextEditor (141, showPreview local),
  LessonResourcesFields (266), LessonStatusPicker (123), LessonEditorActions (110)
Composable useLessonEditor inchangé (showPreview retiré, déplacé dans le rich editor).

Parité: champs en v-model (defineModel) recâblés 1:1 (.number conservé sur les inputs
concernés), mêmes appels services/routes/textes, CSS déplacé verbatim avec @media.
$router.go(-1) conservé via propriété globale. Dette pré-existante conservée: CSS mort
.loading-state/.spinner/@keyframes spin dans l'orchestrateur (template = <ContentLoader>).

Tests: 7 sous-composants (23) + composable (5) + montage vue G5 (1) = verts.
Tous les fichiers ≤300.
